Gestion Erreur OK : Suppression Album

+++ src/main/java/com/myaudiolibrary/web/controller/AlbumController.java
@@ -20,8 +20,7 @@ public class AlbumController {
     @Autowired
     private AlbumRepository albumRepository;
 
-    @Autowired
-    private ArtistRepository artistRepository;
+
 
     //POST http://localhost:5366/albums
     //Ajout d'un album a un artist
@@ -38,10 +37,17 @@ public class AlbumController {
     @ResponseStatus(HttpStatus.NO_CONTENT)//204
     public void deleteAlbum(@PathVariable Long id)
     {
-//        if(albumRepository.findById(id) == null)
-//        {
-//            throw new EntityNotFoundException("Il n'y a pas d'album d'id : " + id);
-//        }
+        //Gestion Erreur : id < 1
+        if (id < 1)
+        {
+            throw new IllegalArgumentException("L'id de l'album doit être égal ou supérieur à 1" );
+        }
+
+        //Gestion Erreur :  404 not found
+        if(albumRepository.findById(id) == null)
+        {
+            throw new EntityNotFoundException("Il n'y a pas d'album d'id : " + id);
+        }
         albumRepository.deleteById(id);
     }
